python/Django/pyproject.toml
[tool.poetry]
name = "django-template"
version = "0.1.0"
description = "Django + GraphQL + Prisma Template Project"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
python = "^3.12"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.0"
django-environ = "^0.11.0"
psycopg2-binary = "^2.9.9"
prisma = "^0.15.0"
djangorestframework = "^3.14.0"
django-filter = "^24.0"
drf-spectacular = "^0.27.0"
graphene-django = "^3.2.0"
graphql-core = "^3.2.0"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"
celery = "^5.4.0"
redis = "^5.0.0"
django-celery-beat = "^2.6.0"
python-dotenv = "^1.0.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
gunicorn = "^23.0.0"
django-debug-toolbar = "^4.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-django = "^4.9.0"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
isort = "^5.13.0"
flake8 = "^7.1.0"
mypy = "^1.11.0"
django-stubs = "^5.1.0"
djangorestframework-stubs = "^3.15.0"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["config", "apps", "core"]
skip = ["migrations"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "-v --tb=short --strict-markers"
testpaths = ["tests"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
